// packages/directus-modules/hook-sync-to-legacy/src/btv.d.ts
export interface components {
  schemas: {
// ...
    ItemsEpisodesTranslations: {
      description?: string | null;
      episodes_id?: number | components["schemas"]["ItemsEpisodes"];
      extra_description?: string | null;
      id?: number;
      is_primary?: boolean;
      languages_code?: string | components["schemas"]["ItemsLanguages"];
      title?: string | null;
    };
// ...
    ItemsSeasonsTranslations: {
      description?: string | null;
      id?: number;
      is_primary?: boolean;
      languages_code?: string | components["schemas"]["ItemsLanguages"];
      legacy_description_id?: number | null;
      legacy_title_id?: number | null;
      seasons_id?: number | components["schemas"]["ItemsSeasons"];
      title?: string | null;
    };
// ...
    ItemsShowsTranslations: {
      description?: string | null;
      id?: number;
      is_primary?: boolean;
      languages_code?: string | components["schemas"]["ItemsLanguages"];
      legacy_description_id?: number | null;
      legacy_tags?: string | null;
      legacy_tags_id?: number | null;
      legacy_title_id?: number | null;
      shows_id?: number | components["schemas"]["ItemsShows"];
      title?: string | null;
    };
// ...
  };
// ...
}